+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.6 - 2024-06-04
+### Changed
+- Moved `FirebaseAccount.restore` to `FirebaseAuth.restoreAccount`
+
+### Deprecated
+- Deprecated `FirebaseAccount.restore` and `FirebaseAccount.apiRestore` in favor of `FirebaseAuth.restoreAccount`
+
+### Fixed
+- Fixed refresh timer not being canceled on `FirebaseAccount.dispose`
+
 ## 2.0.5 - 2024-05-21
 ### Changed
 - Update dependencies